requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
aspy.yaml==1.3.0
attrs==21.4.0
autoflake==1.3.1
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
bcrypt==3.1.7
black==21.12b0
blinker==1.4
Brlapi==0.7.0
cachetools==5.0.0
certifi==2019.11.28
cfgv==3.3.1
chardet==3.0.4
click==8.0.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.4.0
flake8==4.0.1
flasgger==0.9.5
Flask==2.0.2
future==0.18.2
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-bigquery==2.32.0
google-cloud-core==2.2.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.14.0
identify==2.4.0
idna==2.8
importlib-metadata==4.2.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
kazam==1.4.5
keyring==18.0.1
keyrings.alt==3.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
marshmallow==3.14.1
mccabe==0.6.1
mistune==2.0.2
monotonic==1.5
mypy-extensions==0.4.3
netifaces==0.10.4
nodeenv==1.6.0
numpy==1.21.5
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pandas==1.3.5
paramiko==2.6.0
pathspec==0.9.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.4.0
pre-commit==2.0.1
proto-plus==1.19.9
protobuf==3.19.4
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.4.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.8
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
system-service==0.3
systemd-python==234
toml==0.10.2
tomli==1.2.3
typed-ast==1.5.1
typing_extensions==4.0.1
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.10.0
wadllib==1.3.3
Werkzeug==2.0.2
xkit==0.0.0
zipp==3.6.0
zope.interface==4.7.1